+++ projects/js-packages/publicize-components/src/social-store/constants.ts
@@ -1,1 +1,2 @@
 export const SIG_SETTINGS_KEY = 'jetpack_social_image_generator_settings';
+export const UTM_ENABLED_KEY = 'jetpack_social_utm_settings';