+++ projects/js-packages/publicize-components/src/social-store/reducer/site-data.js
@@ -1,5 +1,0 @@
-const siteData = ( state = {} ) => {
-	return state;
-};
-
-export default siteData;