+++ projects/packages/jetpack-mu-wpcom/src/features/help-center/class-wp-rest-help-center-support-interactions.php
@@ -32,6 +32,29 @@ class WP_REST_Help_Center_Support_Interactions extends \WP_REST_Controller {
 				'methods'             => \WP_REST_Server::READABLE,
 				'callback'            => array( $this, 'get_support_interactions' ),
 				'permission_callback' => 'is_user_logged_in',
+				'args'                => array(
+					'status'   => array(
+						'type'     => 'string',
+						'required' => false,
+						'enum'     => array(
+							'open',
+							'resolved',
+						),
+					),
+					'page'     => array(
+						'type'     => 'integer',
+						'required' => false,
+						'default'  => 1,
+						'minimum'  => 1,
+					),
+					'per_page' => array(
+						'type'     => 'integer',
+						'required' => false,
+						'default'  => 10,
+						'minimum'  => 1,
+						'maximum'  => 100,
+					),
+				),
 			)
 		);
 
@@ -54,7 +77,7 @@ class WP_REST_Help_Center_Support_Interactions extends \WP_REST_Controller {
 				'permission_callback' => 'is_user_logged_in',
 				'args'                => array(
 					'event_external_id' => array(
-						'type'     => 'int',
+						'type'     => 'string',
 						'required' => true,
 					),
 					'event_source'      => array(
@@ -78,7 +101,7 @@ class WP_REST_Help_Center_Support_Interactions extends \WP_REST_Controller {
 				'permission_callback' => 'is_user_logged_in',
 				'args'                => array(
 					'event_external_id' => array(
-						'type'     => 'int',
+						'type'     => 'string',
 						'required' => true,
 					),
 					'event_source'      => array(
@@ -92,6 +115,26 @@ class WP_REST_Help_Center_Support_Interactions extends \WP_REST_Controller {
 				),
 			)
 		);
+
+		register_rest_route(
+			$this->namespace,
+			'/' . $this->rest_base . '/(?P<support_interaction_id>[a-zA-Z0-9-]+)/status',
+			array(
+				'methods'             => \WP_REST_Server::EDITABLE,
+				'callback'            => array( $this, 'update_support_interaction_status' ),
+				'permission_callback' => 'is_user_logged_in',
+				'args'                => array(
+					'status' => array(
+						'type'     => 'string',
+						'required' => true,
+						'enum'     => array(
+							'open',
+							'resolved',
+						),
+					),
+				),
+			)
+		);
 	}
 
 	/**
@@ -100,8 +143,13 @@ class WP_REST_Help_Center_Support_Interactions extends \WP_REST_Controller {
 	 * @param \WP_REST_Request $request    The request sent to the API.
 	 */
 	public function get_support_interactions( \WP_REST_Request $request ) {
-		$support_interaction_id = isset( $request['support_interaction_id'] ) ? (int) $request['support_interaction_id'] : null;
-		$body                   = Client::wpcom_json_api_request_as_user( "/support-interactions/$support_interaction_id" );
+		if ( isset( $request['support_interaction_id'] ) ) {
+			$url = "/support-interactions/{$request['support_interaction_id']}";
+		} else {
+			$url = '/support-interactions/?' . http_build_query( stripslashes_deep( $request->get_params() ) );
+		}
+
+		$body = Client::wpcom_json_api_request_as_user( $url );
 
 		if ( is_wp_error( $body ) ) {
 			return $body;
@@ -162,6 +210,34 @@ class WP_REST_Help_Center_Support_Interactions extends \WP_REST_Controller {
 			array(
 				'method' => 'POST',
 				'body'   => $data,
+			)
+		);
+
+		if ( is_wp_error( $body ) ) {
+			return $body;
+		}
+
+		$response = json_decode( wp_remote_retrieve_body( $body ) );
+
+		return rest_ensure_response( $response );
+	}
+
+	/**
+	 * Update support interaction status.
+	 *
+	 * @param \WP_REST_Request $request    The request sent to the API.
+	 */
+	public function update_support_interaction_status( \WP_REST_Request $request ) {
+		$support_interaction_id = isset( $request['support_interaction_id'] ) ? (int) $request['support_interaction_id'] : null;
+
+		$status = $request['status'];
+
+		$body = Client::wpcom_json_api_request_as_user(
+			"/support-interactions/$support_interaction_id/status",
+			'2',
+			array(
+				'method' => 'PUT',
+				'body'   => array( 'status' => $status ),
 			)
 		);