+++ projects/packages/my-jetpack/_inc/components/product-interstitial/index.jsx
@@ -29,10 +29,11 @@ import GoBackLink from '../go-back-link';
 import ProductDetailCard from '../product-detail-card';
 import ProductDetailTable from '../product-detail-table';
 import boostImage from './boost.png';
+import completeImage from './complete.png';
 import crmImage from './crm.png';
 import extrasImage from './extras.png';
 import searchImage from './search.png';
-import socialImage from './social.png';
+import securityImage from './security.png';
 import statsImage from './stats.png';
 import styles from './style.module.scss';
 import videoPressImage from './videopress.png';
@@ -285,6 +286,7 @@ export default function ProductInterstitial( {
 										highlightLastFeature={ highlightLastFeature }
 										isFetching={ isActivating }
 										isFetchingSuccess={ isSuccess }
+										isUpsell={ true }
 									/>
 								) : (
 									children
@@ -349,15 +351,6 @@ export function BoostInterstitial() {
 }
 
 /**
- * CreatorInterstitial component
- *
- * @return {object} CreatorInterstitial react component.
- */
-export function CreatorInterstitial() {
-	return <ProductInterstitial slug="creator" installsPlugin={ true } />;
-}
-
-/**
  * CRMInterstitial component
  *
  * @return {object} CRMInterstitial react component.
@@ -417,17 +410,7 @@ export function ScanInterstitial() {
  * @return {object} SocialInterstitial react component.
  */
 export function SocialInterstitial() {
-	return (
-		<ProductInterstitial slug="social" installsPlugin={ true }>
-			<img
-				src={ socialImage }
-				alt={ __(
-					'Image displaying logos of social media platforms supported by Jetpack Social.',
-					'jetpack-my-jetpack'
-				) }
-			/>
-		</ProductInterstitial>
-	);
+	return <ProductInterstitial slug="social" installsPlugin={ true } bundle="growth" />;
 }
 
 /**
@@ -471,15 +454,8 @@ export function StatsInterstitial() {
 			directCheckout={ true }
 			installsPlugin={ true }
 			ctaButtonLabel={ __( 'Get Stats', 'jetpack-my-jetpack' ) }
-		>
-			<img
-				src={ statsImage }
-				alt={ __(
-					'Illustration showing the Stats feature, highlighting important statistics for your site.',
-					'jetpack-my-jetpack'
-				) }
-			/>
-		</ProductInterstitial>
+			bundle="growth"
+		/>
 	);
 }
 
@@ -492,6 +468,45 @@ export function VideoPressInterstitial() {
 	return (
 		<ProductInterstitial slug="videopress" installsPlugin={ true }>
 			<img src={ videoPressImage } alt="VideoPress" />
+		</ProductInterstitial>
+	);
+}
+
+/**
+ * SecurityInterstitial component
+ *
+ * @return {object} SecurityInterstitial react component.
+ */
+export function SecurityInterstitial() {
+	return (
+		<ProductInterstitial slug="security" installsPlugin={ true }>
+			<img src={ securityImage } alt="Security" />
+		</ProductInterstitial>
+	);
+}
+
+/**
+ * GrowthInterstitial component
+ *
+ * @return {object} GrowthInterstitial react component.
+ */
+export function GrowthInterstitial() {
+	return (
+		<ProductInterstitial slug="growth" installsPlugin={ true }>
+			<img src={ statsImage } alt="Growth" />
+		</ProductInterstitial>
+	);
+}
+
+/**
+ * CompleteInterstitial component
+ *
+ * @return {object} CompleteInterstitial react component.
+ */
+export function CompleteInterstitial() {
+	return (
+		<ProductInterstitial slug="complete" installsPlugin={ true }>
+			<img src={ completeImage } alt="Complete" />
 		</ProductInterstitial>
 	);
 }