+++ projects/packages/my-jetpack/src/products/class-boost.php
@@ -391,4 +391,14 @@ class Boost extends Product {
 
 		return $product_activation;
 	}
+
+	/**
+	 * Return product bundles list
+	 * that supports the product.
+	 *
+	 * @return boolean|array Products bundle list.
+	 */
+	public static function is_upgradable_by_bundle() {
+		return array( 'complete' );
+	}
 }