+++ projects/plugins/boost/app/assets/src/js/features/cornerstone-pages/meta/meta.tsx
@@ -1,7 +1,5 @@
-import { Button, Notice } from '@automattic/jetpack-components';
+import { Button, getRedirectUrl, Notice } from '@automattic/jetpack-components';
 import { __, _n, sprintf } from '@wordpress/i18n';
-import ChevronDown from '$svg/chevron-down';
-import ChevronUp from '$svg/chevron-up';
 import React, { useEffect, useState } from 'react';
 import clsx from 'clsx';
 import styles from './meta.module.scss';
@@ -14,33 +12,33 @@ import { recordBoostEvent } from '$lib/utils/analytics';
 import getSupportLink from '$lib/utils/get-support-link';
 import { useRegenerationReason } from '$features/critical-css/lib/stores/suggest-regenerate';
 import { usePremiumFeatures } from '$lib/stores/premium-features';
-import { useNavigate } from 'react-router-dom';
+import { useRegenerateCriticalCssAction } from '$features/critical-css/lib/stores/critical-css-state';
+import { isSameSiteUrl } from '$lib/utils/is-same-site-url';
+import UpgradeCTA from '$features/upgrade-cta/upgrade-cta';
 
 const Meta = () => {
-	const [ isExpanded, setIsExpanded ] = useState( false );
+	const cornerstonePagesSupportLink = getRedirectUrl( 'jetpack-boost-cornerstone-pages' );
 	const [ cornerstonePages, setCornerstonePages ] = useCornerstonePages();
 	const cornerstonePagesProperties = useCornerstonePagesProperties();
 	const [ { refetch: refetchRegenerationReason } ] = useRegenerationReason();
 	const premiumFeatures = usePremiumFeatures();
-	const isPremium = premiumFeatures.includes( 'support' );
-	const navigate = useNavigate();
-
-	const toggleExpanded = ( newValue: boolean ) => {
-		recordBoostEvent( 'cornerstone_pages_show_options_toggle', {
-			status: newValue ? 'open' : 'close',
-		} );
-		setIsExpanded( newValue );
-	};
+	const isPremium = premiumFeatures.includes( 'cornerstone-10-pages' );
+	const regenerateAction = useRegenerateCriticalCssAction();
 
 	const updateCornerstonePages = ( newValue: string ) => {
 		const newItems = newValue.split( '\n' ).map( line => line.trim() );
 
 		setCornerstonePages( newItems, () => {
 			refetchRegenerationReason();
+			if ( isPremium ) {
+				regenerateAction.mutate();
+			}
 		} );
 	};
 
 	let content = null;
+
+	const listInputRows = isPremium ? 10 : 5;
 
 	if ( cornerstonePagesProperties !== undefined ) {
 		content = (
@@ -48,6 +46,8 @@ const Meta = () => {
 				items={ cornerstonePages.join( '\n' ) }
 				setItems={ updateCornerstonePages }
 				maxItems={ cornerstonePagesProperties.max_pages }
+				defaultValue={ cornerstonePagesProperties.default_pages.join( '\n' ) }
+				inputRows={ listInputRows }
 				description={
 					<>
 						{ createInterpolateElement(
@@ -62,32 +62,6 @@ const Meta = () => {
 							{
 								b: <b />,
 							}
-						) }
-						{ ! isPremium && (
-							<>
-								<br />
-								<br />
-								<b>
-									{ createInterpolateElement(
-										__(
-											'Free users can add only one cornerstone page. <link>Upgrade to add more</link>.',
-											'jetpack-boost'
-										),
-										{
-											link: (
-												// eslint-disable-next-line jsx-a11y/anchor-has-content
-												<a
-													href="#/upgrade"
-													onClick={ () => {
-														recordBoostEvent( 'cornerstone_pages_upgrade_link_clicked', {} );
-														navigate( '/upgrade' );
-													} }
-												/>
-											),
-										}
-									) }
-								</b>
-							</>
 						) }
 					</>
 				}
@@ -127,30 +101,29 @@ const Meta = () => {
 
 	return (
 		<div className={ styles.wrapper } data-testid="cornerstone-pages-meta">
-			<div className={ styles.head }>
-				<div className={ styles.summary }>
-					{ cornerstonePagesProperties &&
-						sprintf(
-							/* translators: %1$d is the number of cornerstone pages added, %2$d is the maximum number allowed */
-							__( '%1$d / %2$d added', 'jetpack-boost' ),
-							cornerstonePages.length,
-							cornerstonePagesProperties.max_pages
-						) }
-				</div>
-				<div className={ styles.actions }>
-					<Button
-						variant="link"
-						size="small"
-						weight="regular"
-						iconSize={ 16 }
-						icon={ isExpanded ? <ChevronUp /> : <ChevronDown /> }
-						onClick={ () => toggleExpanded( ! isExpanded ) }
-					>
-						{ __( 'Show Options', 'jetpack-boost' ) }
-					</Button>
-				</div>
-			</div>
-			{ isExpanded && <div className={ styles.body }>{ content }</div> }
+			<p>
+				{ createInterpolateElement(
+					__(
+						'List the most important pages of your site. These pages will receive specially tailored optimizations, including targeted critical CSS. The Page Speed scores are based on the first cornerstone page. <b><link>Learn More</link></b>',
+						'jetpack-boost'
+					),
+					{
+						link: (
+							// eslint-disable-next-line jsx-a11y/anchor-has-content
+							<a
+								href={ cornerstonePagesSupportLink }
+								target="_blank"
+								rel="noopener noreferrer"
+								onClick={ () => {
+									recordBoostEvent( 'clicked_cornerstone_pages_learn_more', {} );
+								} }
+							/>
+						),
+						b: <b />,
+					}
+				) }
+			</p>
+			<div className={ styles.body }>{ content }</div>
 		</div>
 	);
 };
@@ -160,15 +133,25 @@ type ListProps = {
 	setItems: ( newValue: string ) => void;
 	maxItems: number;
 	description: React.ReactNode | null;
+	defaultValue: string;
+	inputRows?: number;
 };
 
-const List: React.FC< ListProps > = ( { items, setItems, maxItems, description } ) => {
+const List: React.FC< ListProps > = ( {
+	items,
+	setItems,
+	maxItems,
+	description,
+	defaultValue = '',
+	inputRows = 10,
+} ) => {
 	const [ inputValue, setInputValue ] = useState( items );
 	// eslint-disable-next-line @typescript-eslint/no-unused-vars
 	const [ inputInvalid, setInputInvalid ] = useState( false );
 	const [ validationError, setValidationError ] = useState< Error | null >( null );
-
-	const inputRows = Math.min( maxItems, 10 );
+	const premiumFeatures = usePremiumFeatures();
+	const isPremium = premiumFeatures.includes( 'cornerstone-10-pages' );
+	const cornerstonePagesProperties = useCornerstonePagesProperties();
 
 	const validateInputValue = ( value: string ) => {
 		setInputValue( value );
@@ -183,10 +166,11 @@ const List: React.FC< ListProps > = ( { items, setItems, maxItems, description }
 	};
 
 	const validateItems = ( value: string ) => {
-		const lines = value
-			.split( '\n' )
-			.map( line => line.trim() )
-			.filter( line => line.trim() !== '' );
+		const lines = value.split( '\n' ).map( line => line.trim() );
+
+		if ( lines.length === 0 ) {
+			throw new Error( __( 'You must add at least one URL.', 'jetpack-boost' ) );
+		}
 
 		// Check if the number of items exceeds maxItems
 		if ( lines.length > maxItems ) {
@@ -203,6 +187,8 @@ const List: React.FC< ListProps > = ( { items, setItems, maxItems, description }
 			throw new Error( message );
 		}
 
+		const siteUrl = new URL( Jetpack_Boost.site.url );
+
 		for ( const line of lines ) {
 			let url: URL | undefined;
 			try {
@@ -210,10 +196,7 @@ const List: React.FC< ListProps > = ( { items, setItems, maxItems, description }
 			} catch ( e ) {
 				// If the URL is invalid, they have provided a relative URL, which we will allow.
 			}
-			if (
-				url &&
-				url.origin.replace( /\/$/, '' ) !== Jetpack_Boost.site.url.replace( /\/$/, '' )
-			) {
+			if ( url && ! isSameSiteUrl( url, siteUrl ) ) {
 				throw new Error(
 					/* translators: %s is the URL that didn't match the site URL */
 					sprintf( __( 'The URL seems to be a different site: %s', 'jetpack-boost' ), line )
@@ -235,28 +218,56 @@ const List: React.FC< ListProps > = ( { items, setItems, maxItems, description }
 		} );
 	}
 
+	function loadDefaultValue() {
+		validateInputValue( defaultValue );
+		recordBoostEvent( 'cornerstone_pages_load_default', {} );
+	}
+
 	return (
-		<div
-			className={ clsx( styles.section, {
-				[ styles[ 'has-error' ] ]: inputInvalid,
-			} ) }
-		>
-			<textarea
-				value={ inputValue }
-				rows={ inputRows }
-				onChange={ e => validateInputValue( e.target.value ) }
-				id="jb-cornerstone-pages"
-			/>
-			{ inputInvalid && <p className={ styles.error }>{ validationError?.message }</p> }
-			{ description && <div className={ styles.description }>{ description }</div> }
-			<Button
-				disabled={ items === inputValue || inputInvalid }
-				onClick={ save }
-				className={ styles.button }
+		<>
+			<div
+				className={ clsx( styles.section, {
+					[ styles[ 'has-error' ] ]: inputInvalid,
+				} ) }
 			>
-				{ __( 'Save', 'jetpack-boost' ) }
-			</Button>
-		</div>
+				<textarea
+					value={ inputValue }
+					rows={ inputRows }
+					onChange={ e => validateInputValue( e.target.value ) }
+					id="jb-cornerstone-pages"
+				/>
+				{ inputInvalid && <span className={ styles.error }>{ validationError?.message }</span> }
+				{ description && <div className={ styles.description }>{ description }</div> }
+				<Button
+					disabled={ items === inputValue || inputInvalid }
+					onClick={ save }
+					className={ styles.button }
+				>
+					{ __( 'Save', 'jetpack-boost' ) }
+				</Button>
+				<Button
+					disabled={ inputValue === defaultValue }
+					onClick={ loadDefaultValue }
+					className={ styles.button }
+					variant="link"
+				>
+					{ __( 'Load default pages', 'jetpack-boost' ) }
+				</Button>
+			</div>
+			{ ! isPremium && cornerstonePagesProperties && (
+				<div className={ styles.wrapper }>
+					<UpgradeCTA
+						eventName="cornerstone_pages_upgrade_link_clicked"
+						identifier="cornerstone-10-pages"
+						description={ sprintf(
+							/* translators: %d is the number of cornerstone pages. */
+							__( 'Premium users can add up to %d cornerstone pages.', 'jetpack-boost' ),
+							cornerstonePagesProperties.max_pages_premium
+						) }
+					/>
+				</div>
+			) }
+		</>
 	);
 };