+++ projects/plugins/boost/app/lib/Cornerstone_Pages.php
@@ -12,13 +12,22 @@ class Cornerstone_Pages implements Has_Setup {
 	const FREE_MAX_PAGES    = 1;
 
 	public function setup() {
+		$this->register_ds_stores();
+
 		add_filter( 'jetpack_boost_critical_css_providers', array( $this, 'remove_ccss_front_page_provider' ), 10, 2 );
 		add_filter( 'display_post_states', array( $this, 'add_display_post_states' ), 10, 2 );
-		add_action( 'init', array( $this, 'register_ds_stores' ) );
+		add_action( 'init', array( $this, 'set_default_pages' ), 0 );
 	}
 
-	public function register_ds_stores() {
-		$schema = Schema::as_array( Schema::as_string() )->fallback( $this->default_pages() );
+	public function set_default_pages() {
+		$pages = jetpack_boost_ds_get( 'cornerstone_pages_list' );
+		if ( empty( $pages ) ) {
+			jetpack_boost_ds_set( 'cornerstone_pages_list', $this->default_pages() );
+		}
+	}
+
+	private function register_ds_stores() {
+		$schema = Schema::as_array( Schema::as_string() )->fallback( array() );
 		jetpack_boost_register_option( 'cornerstone_pages_list', $schema, new Cornerstone_Pages_Entry( 'cornerstone_pages_list' ) );
 		jetpack_boost_register_readonly_option( 'cornerstone_pages_properties', array( $this, 'get_properties' ) );
 	}
@@ -37,16 +46,17 @@ class Cornerstone_Pages implements Has_Setup {
 
 	private function default_pages() {
 		if ( $this->get_max_pages() === static::FREE_MAX_PAGES ) {
-			return array( '/' );
+			return array( '' );
 		}
 
 		$max_pages               = $this->get_max_pages();
 		$yoast_cornerstone_pages = $this->get_yoast_cornerstone_pages();
 		$woocommerce_pages       = $this->get_woocommerce_pages();
 
-		$homepage = array( '/' );
+		$homepage = array( '' );
 
 		$urls = array_unique( array_merge( $homepage, $woocommerce_pages, $yoast_cornerstone_pages ) );
+		$urls = array_map( 'untrailingslashit', $urls );
 
 		return array_slice( $urls, 0, $max_pages );
 	}
@@ -104,12 +114,23 @@ class Cornerstone_Pages implements Has_Setup {
 	}
 
 	public function get_pages() {
-		return jetpack_boost_ds_get( 'cornerstone_pages_list' );
+		$pages = jetpack_boost_ds_get( 'cornerstone_pages_list' );
+
+		$permalink_structure = get_option( 'permalink_structure' );
+
+		// If permalink structure ends with slash, add trailing slashes
+		if ( $permalink_structure && substr( $permalink_structure, -1 ) === '/' ) {
+			$pages = array_map( 'trailingslashit', $pages );
+		}
+
+		return $pages;
 	}
 
 	public function get_properties() {
 		return array(
-			'max_pages' => $this->get_max_pages(),
+			'max_pages'         => $this->get_max_pages(),
+			'max_pages_premium' => static::PREMIUM_MAX_PAGES,
+			'default_pages'     => array_map( 'home_url', $this->default_pages() ),
 		);
 	}
 
@@ -128,6 +149,6 @@ class Cornerstone_Pages implements Has_Setup {
 	}
 
 	private function get_max_pages() {
-		return Premium_Features::has_any() ? static::PREMIUM_MAX_PAGES : static::FREE_MAX_PAGES;
+		return Premium_Features::has_feature( Premium_Features::CORNERSTONE_TEN_PAGES ) ? static::PREMIUM_MAX_PAGES : static::FREE_MAX_PAGES;
 	}
 }