+++ projects/plugins/boost/app/lib/critical-css/source-providers/Source_Providers.php
@@ -38,12 +38,12 @@ class Source_Providers {
 	 * @var Provider[]
 	 */
 	protected $providers = array(
+		Cornerstone_Provider::class,
 		Post_ID_Provider::class,
 		WP_Core_Provider::class,
 		Singular_Post_Provider::class,
 		Archive_Provider::class,
 		Taxonomy_Provider::class,
-		Cornerstone_Provider::class,
 	);
 
 	public function get_providers() {
@@ -110,19 +110,28 @@ class Source_Providers {
 		return $this->request_cached_css;
 	}
 
+	public function get_current_critical_css_key() {
+		return $this->current_critical_css_key;
+	}
+
 	/**
 	 * Get providers sources.
 	 *
 	 * @return array
 	 */
 	public function get_provider_sources( $context_posts = array() ) {
-		$sources = array();
+		$sources                        = array();
+		$flat_core_and_cornerstone_urls = array();
 
 		$wp_core_provider_urls = WP_Core_Provider::get_critical_source_urls( $context_posts );
-		$flat_wp_core_urls     = array();
 		foreach ( $wp_core_provider_urls as $urls ) {
-			$flat_wp_core_urls = array_merge( $flat_wp_core_urls, $urls );
+			$flat_core_and_cornerstone_urls = array_merge( $flat_core_and_cornerstone_urls, $urls );
 		}
+		$cornerstone_provider_urls = Cornerstone_Provider::get_critical_source_urls( $context_posts );
+		foreach ( $cornerstone_provider_urls as $urls ) {
+			$flat_core_and_cornerstone_urls = array_merge( $flat_core_and_cornerstone_urls, $urls );
+		}
+		$flat_core_and_cornerstone_urls = array_values( array_unique( $flat_core_and_cornerstone_urls ) );
 
 		foreach ( $this->get_providers() as $provider ) {
 			$provider_name = $provider::get_provider_name();
@@ -134,10 +143,10 @@ class Source_Providers {
 					continue;
 				}
 
-				// This removes the home and blog pages from the list of pages,
+				// This removes core and cornerstone URLs from the list of URLs,
 				// so they don't belong to two separate groups.
 				if ( ! in_array( $provider, array( WP_Core_Provider::class, Cornerstone_Provider::class ), true ) ) {
-					$urls = array_values( array_diff( $urls, $flat_wp_core_urls ) );
+					$urls = array_values( array_diff( $urls, $flat_core_and_cornerstone_urls ) );
 				}
 
 				if ( empty( $urls ) ) {
@@ -170,7 +179,7 @@ class Source_Providers {
 		 * Filters the list of Critical CSS source providers.
 		 *
 		 * @param array $sources The list of Critical CSS source providers.
-		 * @since $$next-version$$
+		 * @since 3.6.0
 		 */
 		return apply_filters( 'jetpack_boost_critical_css_providers', $sources );
 	}