+++ projects/plugins/jetpack/tests/jest.config.extensions.js
@@ -3,7 +3,6 @@ const baseConfig = require( './jest.config.base.js' );
 module.exports = {
 	...baseConfig,
 	roots: [ '<rootDir>/extensions/' ],
-	coverageDirectory: 'coverage/extensions',
 	setupFiles: [ ...baseConfig.setupFiles, '<rootDir>/tests/jest-globals.extensions.js' ],
 	testPathIgnorePatterns: [
 		...baseConfig.testPathIgnorePatterns,
@@ -13,4 +12,9 @@ module.exports = {
 		...baseConfig.moduleNameMapper,
 		'\\.(css|less|sass|scss)$': '<rootDir>/tests/styles-mock.js',
 	},
+	coverageDirectory: baseConfig.coverageDirectory + '/extensions',
+	collectCoverageFrom: [
+		'<rootDir>/extensions/**/*.{js,mjs,cjs,jsx,ts,tsx,mts,cts}',
+		...baseConfig.collectCoverageFrom.slice( 3 ),
+	],
 };