+++ projects/plugins/jetpack/tests/jest.config.gui.js
@@ -6,5 +6,11 @@ module.exports = {
 	testMatch: [ '<rootDir>/_inc/client/test/main.js', '<rootDir>/_inc/client/**/test/component.js' ],
 	setupFilesAfterEnv: [ ...baseConfig.setupFilesAfterEnv, '<rootDir>/tests/jest-globals.gui.js' ],
 	transformIgnorePatterns: [ '/node_modules/(?!(.pnpm|@automattic)/)' ],
-	coverageDirectory: 'coverage/gui',
+	coverageDirectory: baseConfig.coverageDirectory + '/gui',
+	collectCoverageFrom: [
+		'<rootDir>/_inc/**/*.{js,mjs,cjs,jsx,ts,tsx,mts,cts}',
+		'<rootDir>/3rd-party/**/*.{js,mjs,cjs,jsx,ts,tsx,mts,cts}',
+		'<rootDir>/views/**/*.{js,mjs,cjs,jsx,ts,tsx,mts,cts}',
+		...baseConfig.collectCoverageFrom,
+	],
 };