+++ projects/packages/explat/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2024-11-14
+### Removed
+- General: Update minimum PHP version to 7.2. [#40147]
+
+## [0.1.15] - 2024-11-11
+### Changed
+- Updated package dependencies. [#39999] [#40000] [#40060]
+
+## [0.1.14] - 2024-11-04
+### Added
+- Enable test coverage. [#39961]
+
 ## [0.1.13] - 2024-10-29
 ### Changed
 - Internal updates.
@@ -15,8 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.11] - 2024-10-10
 ### Changed
-- Updated package dependencies. [#39649]
-- Updated package dependencies. [#39707]
+- Updated package dependencies. [#39649] [#39707]
 
 ## [0.1.10] - 2024-10-07
 ### Changed
@@ -68,6 +79,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ExPlat: add condition to prevent fetching the experiment assignment if there's not anon id (meaning that Tracks is likely disabled) [#38327]
 - Updated package dependencies. [#38132]
 
+[0.2.0]: https://github.com/Automattic/jetpack-explat/compare/v0.1.15...v0.2.0
+[0.1.15]: https://github.com/Automattic/jetpack-explat/compare/v0.1.14...v0.1.15
+[0.1.14]: https://github.com/Automattic/jetpack-explat/compare/v0.1.13...v0.1.14
 [0.1.13]: https://github.com/Automattic/jetpack-explat/compare/v0.1.12...v0.1.13
 [0.1.12]: https://github.com/Automattic/jetpack-explat/compare/v0.1.11...v0.1.12
 [0.1.11]: https://github.com/Automattic/jetpack-explat/compare/v0.1.10...v0.1.11